remove NodeUpgradeJob controller local map cache
refactor NodeUpgradeJob: using FilterResourceEventHandler directly
